refactor: 이름 속성만 가지기에 Name -> Player 역할 수정
